+++ Project/Engine/CAnimationState.h
@@ -16,26 +16,8 @@ public:
 	void SetSpeed(float _fSpeed) { m_fSpeed = _fSpeed; }
 	
 	vector<tAnimationKeyFrame>& GetBoneTransforms();
-	void SetTickByPercent();
 public:
 	CAnimationState();
 	CAnimationState(const CAnimationState& _other);
 	~CAnimationState();
-};
-
-class CAnimationTransition
-{
-private:
-	CAnimationState* m_pPrevState;
-	CAnimationState* m_pNextState;
-	
-	bool m_bHasExitTime;
-	double m_dExitTime; // percent
-	bool m_bFixedDuration;
-	double m_dTransitionDuration;
-	double m_dTransitionOffset;
-
-public:
-	CAnimationTransition();
-	~CAnimationTransition();
 };
