+++ Project/Engine/CAnimationStateMachine.cpp
@@ -1,0 +1,10 @@
+#include "pch.h"
+#include "CAnimationStateMachine.h"
+
+CAnimationStateMachine::CAnimationStateMachine(CAnimator3D* _pAnimator)
+{
+}
+
+CAnimationStateMachine::~CAnimationStateMachine()
+{
+}
