+++ Project/Engine/CAnimationStateMachine.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "CAnimator3D.h"
+class CAnimationStateMachine
+{
+private:
+public:
+	CAnimationStateMachine(CAnimator3D* _pAnimator);
+	~CAnimationStateMachine();
+};
+
