+++ Project/Engine/CAnimationTransition.cpp
@@ -1,0 +1,13 @@
+#include "pch.h"
+#include "CAnimationTransition.h"
+#include "CAnimationState.h"
+
+CAnimationTransition::CAnimationTransition()
+	: m_pPrevState(nullptr)
+	, m_pNextState(nullptr)
+{
+}
+
+CAnimationTransition::~CAnimationTransition()
+{
+}
