+++ Project/Engine/Engine.vcxproj
@@ -208,6 +208,8 @@ call EngineCopy.bat</Command>
     <ClInclude Include="CAnim2D.h" />
     <ClInclude Include="CAnimationClip.h" />
     <ClInclude Include="CAnimationState.h" />
+    <ClInclude Include="CAnimationStateMachine.h" />
+    <ClInclude Include="CAnimationTransition.h" />
     <ClInclude Include="CAnimator2D.h" />
     <ClInclude Include="CAnimator3D.h" />
     <ClInclude Include="CBoneHolder.h" />
@@ -274,6 +276,8 @@ call EngineCopy.bat</Command>
     <ClCompile Include="CAnim2D.cpp" />
     <ClCompile Include="CAnimationClip.cpp" />
     <ClCompile Include="CAnimationState.cpp" />
+    <ClCompile Include="CAnimationStateMachine.cpp" />
+    <ClCompile Include="CAnimationTransition.cpp" />
     <ClCompile Include="CAnimator2D.cpp" />
     <ClCompile Include="CAnimator3D.cpp" />
     <ClCompile Include="CBoneHolder.cpp" />
