+++ Project/Engine/Engine.vcxproj.filters
@@ -196,6 +196,12 @@
     <ClInclude Include="CCollider3D.h">
       <Filter>Source\06. Component\2. Collider</Filter>
     </ClInclude>
+    <ClInclude Include="CAnimationStateMachine.h">
+      <Filter>Source\06. Component\6. Animator\Animator3D\AnimationState</Filter>
+    </ClInclude>
+    <ClInclude Include="CAnimationTransition.h">
+      <Filter>Source\06. Component\6. Animator\Animator3D\AnimationState</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Engine.cpp">
@@ -380,6 +386,12 @@
     </ClCompile>
     <ClCompile Include="CCollider3D.cpp">
       <Filter>Source\06. Component\2. Collider</Filter>
+    </ClCompile>
+    <ClCompile Include="CAnimationStateMachine.cpp">
+      <Filter>Source\06. Component\6. Animator\Animator3D\AnimationState</Filter>
+    </ClCompile>
+    <ClCompile Include="CAnimationTransition.cpp">
+      <Filter>Source\06. Component\6. Animator\Animator3D\AnimationState</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
